Opt cURL requests into 1.2
There seems to be some combinations of systems and libraries that given
the option for a range in 1.0 to 1.2 will choose TLS 1.0 or 1.1 and then
have their requests fail once they go to the Stripe API. This change
opts everyone into TLS 1.2, thus avoiding the problem.

There is a possibility that this could affect the upgrades for users on
versions of cURL that are so old that they don't have access to 1.2, but
we'll need to help these people upgrade anyway given that the
depreaction dates for obsolete version of TLS are in the not-too-distant
future.