Fixes tests on master by using multiple plans
As described in #315, it seems that some server side logic changed so
that multiple plans of the same ID under the same subscription are no
longer allowed. A couple tests we had depended on that behavior though,
and so when that changed, master broke.

Here we use a similar testing strategy, but just use a new plan when we
need to do an update.

Fixes #315.