Backport setAppInfo to legacy version of the bindings
We have a lot of third-party plugins/platforms that use the legacy version
The main reason is to support old PHP installations. We still want them to
explicitly showcase who they are so that we can better help their users.

This is a first attempt at bringing the functionality in for a potential
release as 1.18.1