Clean up + sort `$nestedUpdateableAttributes` array
Here we cleanup `$nestedUpdatableAttributes` in `StripeObject` by moving
every field name to its own line and then alphabetizing them. This makes
it a little easier to see what's going on when people add new ones, and
gives them a deterministic place in the array with which to make the
addition.