Add PHPDoc return types
Symfony PHPUnit Bridge has a deprecation helper that warns when things might break in the future.

It currently reports the following errors when using Stripe PHP SDK:

```
Method "ArrayAccess::offsetExists()" might add "bool" as a native return type declaration in the future. Do the same in implementation "Stripe\StripeObject" now to avoid errors or add an explicit @return annotation to suppress this message.
Method "ArrayAccess::offsetGet()" might add "mixed" as a native return type declaration in the future. Do the same in implementation "Stripe\StripeObject" now to avoid errors or add an explicit @return annotation to suppress this message.
Method "ArrayAccess::offsetSet()" might add "void" as a native return type declaration in the future. Do the same in implementation "Stripe\StripeObject" now to avoid errors or add an explicit @return annotation to suppress this message.
Method "ArrayAccess::offsetUnset()" might add "void" as a native return type declaration in the future. Do the same in implementation "Stripe\StripeObject" now to avoid errors or add an explicit @return annotation to suppress this message.
Method "Countable::count()" might add "int" as a native return type declaration in the future. Do the same in implementation "Stripe\StripeObject" now to avoid errors or add an explicit @return annotation to suppress this message.
Method "JsonSerializable::jsonSerialize()" might add "mixed" as a native return type declaration in the future. Do the same in implementation "Stripe\StripeObject" now to avoid errors or add an explicit @return annotation to suppress this message.
```

By adding these PHPDoc types, the problem disappears. It's also more clear for users of the package.

Fixes #1210